requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asn1crypto==1.5.1
bcrypt==4.0.1
blinker==1.6.2
build==1.0.3
cffi==1.15.1
click==8.1.6
cryptography==41.0.3
distlib==0.3.7
exceptiongroup==1.2.2
fastapi==0.112.0
filelock==3.12.3
Flask==3.0.0
Flask-Cors==4.0.0
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
packaging==23.1
pg8000==1.30.3
platformdirs==3.10.0
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
PyMySQL==1.1.0
pyOpenSSL==23.2.0
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
scramp==1.4.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.19
starlette==0.37.2
tomli==2.0.1
typing_extensions==4.12.2
uvicorn==0.30.6
virtualenv==20.24.4
Werkzeug==3.0.1
